refactor: correct package name
