+++ classfile-fingerprint/src/main/java/io/github/algomaster99/data/Fingerprint.java
@@ -1,3 +1,3 @@
-package org.example.data;
+package io.github.algomaster99.data;
 
 public record Fingerprint(String groupId, String artifactId, String version, String className, String hash) {}
