+++ classfile-fingerprint/src/test/resources-its/io/github/algomaster99/it/GenerateMojoIT/guava/src/test/resources/expected-classfile.sha256.jsonl
