+++ classfile-fingerprint/src/test/resources-its/io/github/algomaster99/it/GenerateMojoIT/multi_module/pom.xml
