+++ classfile-fingerprint/src/test/resources-its/io/github/algomaster99/it/GenerateMojoIT/native_dependency/pom.xml
