+++ watchdog-agent/pom.xml
@@ -49,7 +49,7 @@
                     <transformers>
                         <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
                             <manifestEntries>
-                                <Premain-Class>org.example.Terminator</Premain-Class>
+                                <Premain-Class>io.github.algomaster99.Terminator</Premain-Class>
                                 <Can-Retransform-Classes>true</Can-Retransform-Classes>
                                 <Multi-Release>true</Multi-Release>
                             </manifestEntries>
