+++ watchdog-agent/src/main/java/io/github/algomaster99/ByteBuddyHelper.java
@@ -1,4 +1,4 @@
-package org.example;
+package io.github.algomaster99;
 
 import net.bytebuddy.ClassFileVersion;
 import net.bytebuddy.dynamic.TargetType;
