+++ watchdog-agent/src/main/java/io/github/algomaster99/Terminator.java
@@ -1,4 +1,4 @@
-package org.example;
+package io.github.algomaster99;
 
 import java.io.File;
 import java.io.IOException;
